Перенесена реализация геттеров в AbstractAnimal. Файлы разбиты на пакеты. И Predator и Prey теперь абстрактны.
