--- src/main/java/ru/mts/hw_3/AbstractAnimal.java
@@ -1,10 +1,0 @@
-package ru.mts.hw_3;
-
-import java.math.BigDecimal;
-
-public abstract class AbstractAnimal implements Animal {
-    protected String name;
-    protected String breed;
-    protected String character;
-    protected BigDecimal cost;
-}
